+++ backend-services/user-service/src/main/java/com/ai/qa/user/application/userService.java
@@ -1,73 +1,19 @@
 package com.ai.qa.user.application;
 
-import com.ai.qa.user.common.CommonUtil;
 import com.ai.qa.user.domain.entity.User;
-import com.ai.qa.user.infrastructure.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import javax.transaction.Transactional;
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
-@Service
-public class userService {
+public interface UserService {
     
-    @Autowired
-    private UserRepository userRepository;
+    User login(String username, String pwd);
     
-    public User login(String username, String pwd) {        
-        User user = userRepository.findByUsername(username);
-        String encryptedPassword = CommonUtil.encryptPassword(pwd);
-        if (user != null && user.getPassword().equals(encryptedPassword)) {
-            return user;
-        }
-        return null;
-    }
-
-    @Transactional
-    public User register(String username, String pwd) {
-        return register(username, pwd, null);
-    }
+    User register(String username, String pwd);
     
-    @Transactional
-    public User register(String username, String pwd, String nick) {
-        
-        // 检查用户是否已存在
-        User existingUser = userRepository.findByUsername(username);
-        if (existingUser != null) {
-            return null; // 用户已存在
-        }
-        
-        User user = new User();
-        user.setUsername(username);
-        user.setPassword(CommonUtil.encryptPassword(pwd)); 
-        user.setNick(nick);
-        
-        // 保存到数据库
-        User savedUser = userRepository.save(user);
-        return savedUser;
-    }
+    User register(String username, String pwd, String nick);
     
-    @Transactional
-    public boolean updateNick(String nick, Long userId) {
-        // 实际项目中需要实现更新用户昵称逻辑
-        Optional<User> userOptional = userRepository.findById(userId);
-        if (userOptional.isPresent()) {
-            User user = userOptional.get();
-            user.setNick(nick);
-            userRepository.save(user);
-            return true;
-        }
-        return false;
-    }
+    boolean updateNick(String nick, Long userId);
     
-    public User getUserById(Long userId) {
-        Optional<User> userOptional = userRepository.findById(userId);
-        return userOptional.orElse(null);
-    }
+    User getUserById(Long userId);
     
-    public List<User> getUsers() {
-        return userRepository.findAll();
-    }
+    List<User> getUsers();
 }
